Update versions of packages in Ubuntu 20.04 images
This commit updates the versions of various packages in the Ubuntu 20.04
images. It affects specifically the plain image and the CUDA image; the
OneAPI image is not affected. The changes are as follows.

1. The primary goal is to upgrade CMake from version 3.17.3 to version
   3.21.2, because this will allow us to build CUDA code with nvcc while
   adhering to the C++17 standard (see acts-project/traccc#78).
2. Version 6.20.04 of ROOT does not compile with CMake 3.21.2 because it
   fails to import some Python targets. In order to remedy this (and to
   look towards the future) I have bumped the ROOT version to 6.24.06.
3. The new version of ROOT requires nlohmann's JSON library to be
   installed, which I have added to the images accordingly.
4. Finally, DD4hep v01-11-02 cannot be compiled against the new ROOT
   version, and such I have bumped DD4hep up to v01-17.

In addition, I have fixed a mistake in the comments which Attila found
very annoying.

I have built both of these images on my local machine, and the build
seems to work well. I am hopeful that they will build succesfully in the
CI as well!