Add Ubuntu 24.04 machine with LLVM 19
LLVM 19 adds the new (and very conveniently named)
`-Wmissing-template-arg-list-after-template-kw` warning flag which fires
a lot on the existing ACTS code. The new version of OneAPI is based on
this version of LLVM, which is causing us a bit of grief in e.g.
https://github.com/acts-project/traccc/pull/793. In order to resolve
this issue I want to add an LLVM 19 build to ACTS.